requirements.txt
arrow==0.13.0
Babel==2.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
dash==0.36.0
dash-core-components==0.43.0
dash-html-components==0.13.5
dash-renderer==0.17.0
dash-resumable-upload==0.0.3
decorator==4.3.0
Flask==1.0.2
Flask-Compress==1.4.0
gunicorn==19.9.0
humanize==0.5.1
idna==2.7
inflect==0.3.1
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.0
nbformat==4.4.0
numpy==1.14.5
pandas==0.23.3
plotly==3.6.0
python-dateutil==2.7.3
python-dotenv==0.10.1
pytz==2018.5
redis==2.10.6
requests==2.21.0
retrying==1.3.3
rq==0.12.0
-e git+https://github.com/Miketsukami/rq-win.git@2040c4c86237108173fd018af530f167fb00345e#egg=rq_win
six==1.11.0
times==0.7
tqdm==4.23.4
traitlets==4.3.2
urllib3==1.23
Werkzeug==0.14.1
xlrd==1.1.0
XlsxWriter==1.1.2
